# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "",
        "contact": {
            "email": "afriantomahmud@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/payment-method": {
            "get": {
                "tags": [
                    "Paymentmethod"
                ],
                "summary": "Get a list of Payment Method",
                "description": "Returns a list of Payment Method",
                "operationId": "9f872ab15ef155437b7cdb9d8af13874",
                "responses": {
                    "200": {
                        "description": "Successful response: List of Payment Method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PaymentMethodResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Paymentmethod"
                ],
                "summary": "Store a new Payment Method",
                "description": "Store a new Payment Method in database",
                "operationId": "e6e20cce545435c81b06284caa2c68b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentMethodRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response: object of Payment Method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaymentMethodResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payment-method/{id}": {
            "get": {
                "tags": [
                    "Paymentmethod"
                ],
                "summary": "Get a single of Payment Method",
                "description": "Returns a single of Payment Method by ID",
                "operationId": "1e14523f3a33e1274c03d85c9e1be646",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Payment Method to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response: object of payment method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaymentMethodResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Paymentmethod"
                ],
                "summary": "Update a new Payment Method",
                "description": "Update a new Payment Method in database",
                "operationId": "4787c8e2d395088048273db2652a8e29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Payment Method to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PaymentMethodRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response: object of payment method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PaymentMethodResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Paymentmethod"
                ],
                "summary": "Delete a new Payment Method",
                "description": "Delete a new Payment Method in database",
                "operationId": "4478b36c5c5ac7aebd95743c54385a99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Payment Method to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Data Units"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PaymentMethodRequest": {
                "title": "Payment Method Request",
                "description": "Request structure for payment methods",
                "required": [
                    "name",
                    "is_default"
                ],
                "properties": {
                    "name": {
                        "description": "name",
                        "type": "string"
                    },
                    "is_default": {
                        "description": "is_default",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "PaymentMethodResource": {
                "title": "Payment Method Resource",
                "description": "Payment Method Resource",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the unit",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the payment method",
                        "type": "string"
                    },
                    "is_default": {
                        "description": "The is_default of the payment method",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}